+++ src/trpc/server/middleware.ts
@@ -1,23 +1,23 @@
+import { Params } from './Params';
 import { ProcedureType } from './procedure';
 import { MiddlewareMarker } from './utils';
 
 /**
  * @internal
  */
-interface MiddlewareResultBase<TParams extends Params> {
+interface MiddlewareResultBase {
   /**
    * All middlewares should pass through their `next()`'s output.
    * Requiring this marker makes sure that can't be forgotten at compile-time.
    */
   readonly marker: MiddlewareMarker;
-  ctx: TParams['ctx'];
 }
 
 /**
  * @internal
  */
-interface MiddlewareOKResult<TParams extends Params>
-  extends MiddlewareResultBase<TParams> {
+interface MiddlewareOKResult<_TParams extends Params>
+  extends MiddlewareResultBase {
   ok: true;
   data: unknown;
   // this could be extended with `input`/`rawInput` later
@@ -26,45 +26,19 @@ interface MiddlewareOKResult<TParams extends Params>
 /**
  * @internal
  */
-interface MiddlewareErrorResult<TParams extends Params>
-  extends MiddlewareResultBase<TParams> {
+interface MiddlewareErrorResult<_TParams extends Params>
+  extends MiddlewareResultBase {
   ok: false;
   error: Error;
   // we could guarantee it's always of this type
 }
 
-export type MiddlewareResult<TParams extends Params> =
-  | MiddlewareOKResult<TParams>
-  | MiddlewareErrorResult<TParams>;
-
 /**
  * @internal
  */
-export interface Params<
-  TContext = unknown,
-  TInputIn = unknown,
-  TInputOut = unknown,
-  TOutputIn = unknown,
-  TOutputOut = unknown,
-> {
-  ctx: TContext;
-  /**
-   * @internal
-   */
-  _output_in: TOutputIn;
-  /**
-   * @internal
-   */
-  _output_out: TOutputOut;
-  /**
-   * @internal
-   */
-  _input_in: TInputIn;
-  /**
-   * @internal
-   */
-  _input_out: TInputOut;
-}
+export type MiddlewareResult<TParams extends Params> =
+  | MiddlewareOKResult<TParams>
+  | MiddlewareErrorResult<TParams>;
 
 /**
  * @internal
@@ -73,7 +47,7 @@ export type MiddlewareFunction<
   TParams extends Params,
   TParamsAfter extends Params,
 > = (opts: {
-  ctx: TParams['ctx'];
+  ctx: TParams['_ctx_out'];
   type: ProcedureType;
   path: string;
   input: TParams['_input_out'];
@@ -82,7 +56,8 @@ export type MiddlewareFunction<
     (): Promise<MiddlewareResult<TParams>>;
     <$TContext>(opts: { ctx: $TContext }): Promise<
       MiddlewareResult<{
-        ctx: $TContext;
+        _ctx_in: TParams['_ctx_in'];
+        _ctx_out: $TContext;
         _input_in: TParams['_input_in'];
         _input_out: TParams['_input_out'];
         _output_in: TParams['_output_in'];
@@ -92,12 +67,6 @@ export type MiddlewareFunction<
   };
 }) => Promise<MiddlewareResult<TParamsAfter>>;
 
-export type inferMiddlewareParams<
-  TMiddleware extends MiddlewareFunction<any, any>,
-> = TMiddleware extends MiddlewareFunction<any, infer $TParams>
-  ? $TParams
-  : never;
-
 /**
  * @internal
  */
@@ -105,7 +74,8 @@ export function createMiddlewareFactory<TContext>() {
   return function createMiddleware<$TNewParams extends Params>(
     fn: MiddlewareFunction<
       {
-        ctx: TContext;
+        _ctx_in: TContext;
+        _ctx_out: TContext;
         _input_out: unknown;
         _input_in: unknown;
         _output_in: unknown;
